# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mencatat servis kendaraan dan manajemen foto per transaksi.",
        "title": "Service Kendaraan API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/kendaraan": {
            "get": {
                "description": "Mengambil data kendaraan. Contoh: ?page=1\u0026limit=5 atau ?page=2\u0026limit=15",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Kendaraan"
                ],
                "summary": "Ambil list kendaraan dengan pagination fleksibel",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman ke berapa (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman (bisa 5, 10, 15, dll)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Aktif (true, false, all)",
                        "name": "aktif",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "nama",
                        "description": "Kolom untuk sorting (nama, manufaktur, created_at)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "asc",
                        "description": "Arah urutan (asc, desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cari berdasarkan nama kendaraan",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.PaginationResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Kendaraan"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Kendaraan"
                ],
                "summary": "Tambah Kendaraan Baru",
                "parameters": [
                    {
                        "description": "Data Kendaraan (Enum Jenis: Mobil, Motor)",
                        "name": "kendaraan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.KendaraanDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Kendaraan"
                        }
                    }
                }
            }
        },
        "/api/kendaraan/getKendaraanByStatus/{status}": {
            "get": {
                "description": "Mengambil data Kendaraan berdasarkan status. Contoh: /api/kendaraan/getKendaraanByStatus/",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Kendaraan"
                ],
                "summary": "Ambil Data Kendaraan berdasarkan status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Kendaraan"
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/kendaraan/{no_polisi}": {
            "get": {
                "description": "Mengambil data spesifik kendaraan berdasarkan ID. Contoh: /api/kendaraan/G1234",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Kendaraan"
                ],
                "summary": "Ambil Data Kendaraan berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "No Polisi Kendaraan",
                        "name": "no_polisi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Kendaraan"
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Mengubah data Kendaraan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Kendaraan"
                ],
                "summary": "Update Data Kendaraan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "No Polisi",
                        "name": "no_polisi",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Kendaraan",
                        "name": "kendaraan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Kendaraan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Kendaraan"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.KendaraanDTO": {
            "type": "object",
            "required": [
                "jenis",
                "manufaktur",
                "nama",
                "no_polisi",
                "tahun",
                "warna"
            ],
            "properties": {
                "jenis": {
                    "type": "string",
                    "enum": [
                        "Mobil",
                        "Motor"
                    ]
                },
                "manufaktur": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_polisi": {
                    "type": "string"
                },
                "tahun": {
                    "type": "string"
                },
                "warna": {
                    "type": "string"
                }
            }
        },
...
        "models.Kendaraan": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "jenis": {
                    "type": "string"
                },
                "manufaktur": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_polisi": {
                    "type": "string"
                },
                "tahun": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "warna": {
                    "type": "string"
                }
            }
        },
...
        "utils.PaginationResult": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_data": {
                    "type": "integer",
                    "example": 100
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        }
    }
}